requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.7
    # via requests
idna==3.11
    # via requests
numpy==2.4.4
    # via pandas
pandas==3.0.2
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via -r requirements.in
requests==2.33.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
urllib3==2.6.3
    # via requests
